+++ stroom-core-client/src/main/java/stroom/entity/client/presenter/MarkdownEditPresenter.java
@@ -17,16 +17,27 @@
 
 package stroom.entity.client.presenter;
 
+import stroom.alert.client.event.AlertEvent;
 import stroom.document.client.event.DirtyEvent;
 import stroom.document.client.event.DirtyEvent.DirtyHandler;
 import stroom.document.client.event.HasDirtyHandlers;
+import stroom.editor.client.presenter.ChangeThemeEvent;
 import stroom.editor.client.presenter.EditorPresenter;
-import stroom.editor.client.presenter.HtmlPresenter;
 import stroom.entity.client.presenter.MarkdownEditPresenter.MarkdownEditView;
+import stroom.iframe.client.presenter.IFramePresenter;
+import stroom.iframe.client.presenter.IFramePresenter.SandboxOption;
+import stroom.preferences.client.UserPreferencesManager;
 import stroom.svg.client.SvgImages;
+import stroom.svg.client.SvgPresets;
+import stroom.ui.config.client.UiConfigCache;
+import stroom.ui.config.shared.Themes.ThemeType;
 import stroom.widget.button.client.ButtonPanel;
+import stroom.widget.button.client.ButtonView;
 import stroom.widget.button.client.InlineSvgToggleButton;
+import stroom.widget.util.client.MouseUtil;
+import stroom.widget.util.client.SafeHtmlUtil;
 
+import com.google.gwt.user.client.Window;
 import com.google.gwt.user.client.ui.Widget;
 import com.google.inject.Inject;
 import com.google.web.bindery.event.shared.EventBus;
@@ -42,32 +53,54 @@ public class MarkdownEditPresenter
         extends MyPresenterWidget<MarkdownEditView>
         implements HasDirtyHandlers, HasToolbar {
 
+    // See prismjs-light.css which is a modified version of a prismjs light theme with .stroom-light-theme
+    // selectivity added into to stop it conflicting with the dark theme.
+    private static final String PRISM_LIGHT_THEME_CLASS_NAME = "stroom-light-theme";
+    private static final String MARKDOWN_FRAME_ID = "markdown-frame";
+
     private final EditorPresenter codePresenter;
-    private final HtmlPresenter htmlPresenter;
+    private final IFramePresenter iFramePresenter;
+    private final UiConfigCache uiConfigCache;
     private final InlineSvgToggleButton editModeButton;
+    private final ButtonView helpButton;
+    private final ButtonPanel toolbar;
+    private final UserPreferencesManager userPreferencesManager;
     private boolean reading;
     private boolean readOnly = true;
     private boolean editMode;
-    private final ButtonPanel toolbar;
 
     @Inject
     public MarkdownEditPresenter(final EventBus eventBus,
                                  final MarkdownEditView view,
                                  final EditorPresenter editorPresenter,
-                                 final HtmlPresenter htmlPresenter) {
+                                 final IFramePresenter iFramePresenter,
+                                 final UiConfigCache uiConfigCache,
+                                 final UserPreferencesManager userPreferencesManager) {
         super(eventBus, view);
         this.codePresenter = editorPresenter;
-        this.htmlPresenter = htmlPresenter;
+        this.iFramePresenter = iFramePresenter;
+        this.uiConfigCache = uiConfigCache;
+        this.userPreferencesManager = userPreferencesManager;
         codePresenter.setMode(AceEditorMode.MARKDOWN);
-        view.setView(htmlPresenter.getView());
+
+        // Markdown is mostly wordy content so wrap long lines to make it easier for the user.
+        codePresenter.getLineWrapOption().setOn();
+
+        iFramePresenter.setId(MARKDOWN_FRAME_ID);
+        iFramePresenter.setSandboxEnabled(true, SandboxOption.ALLOW_POPUPS);
+        view.setView(iFramePresenter.getView());
 
         editModeButton = new InlineSvgToggleButton();
-        editModeButton.setSvg(SvgImages.EDIT);
+        editModeButton.setSvg(SvgImages.MONO_EDIT);
         editModeButton.setTitle("Edit");
         editModeButton.setEnabled(true);
 
         toolbar = new ButtonPanel();
         toolbar.addButton(editModeButton);
+        helpButton = toolbar.addButton(SvgPresets.HELP.title("Documentation help"));
+
+        registerHandler(eventBus.addHandler(ChangeThemeEvent.getType(), event ->
+                updateMarkdownOnIFramePresenter()));
     }
 
     @Override
@@ -89,9 +122,15 @@ public class MarkdownEditPresenter
                 if (editMode) {
                     getView().setView(codePresenter.getView());
                 } else {
-                    htmlPresenter.setHtml(getHtml(codePresenter.getText()));
-                    getView().setView(htmlPresenter.getView());
+                    // Raw markdown likely been changed to update the iframe cotent
+                    updateMarkdownOnIFramePresenter();
+                    getView().setView(iFramePresenter.getView());
                 }
+            }
+        }));
+        registerHandler(helpButton.addClickHandler(e -> {
+            if (MouseUtil.isPrimary(e)) {
+                showHelp();
             }
         }));
     }
@@ -111,26 +150,119 @@ public class MarkdownEditPresenter
         return codePresenter.getText();
     }
 
-    public void setText(String text) {
-        if (text == null) {
-            text = "";
+    public void setText(String rawMarkdown) {
+        if (rawMarkdown == null) {
+            rawMarkdown = "";
         }
 
         reading = true;
-        codePresenter.setText(text);
+        codePresenter.setText(rawMarkdown);
         reading = false;
+        updateMarkdownOnIFramePresenter();
+    }
 
-        htmlPresenter.setHtml(getHtml(text));
+    private void showHelp() {
+        uiConfigCache.get()
+                .onSuccess(result -> {
+                    final String helpUrl = result.getHelpUrlDocumentation();
+                    if (helpUrl != null && helpUrl.trim().length() > 0) {
+                        Window.open(helpUrl, "_blank", "");
+                    } else {
+                        AlertEvent.fireError(
+                                MarkdownEditPresenter.this,
+                                "Help is not configured!",
+                                null);
+                    }
+                })
+                .onFailure(caught -> AlertEvent.fireError(
+                        MarkdownEditPresenter.this,
+                        caught.getMessage(),
+                        null));
+    }
+
+    private void updateMarkdownOnIFramePresenter() {
+        final String rawMarkdown = codePresenter.getText();
+        final String markdownAsHtml = convertMarkdownToHtml(rawMarkdown == null
+                ? ""
+                : rawMarkdown);
+
+        final String currentPreferenceClasses = userPreferencesManager.getCurrentPreferenceClasses();
+        final String prismThemeClassName = getPrismThemeClassName();
+
+        // No point using SafeHtmlBuilder as we have no choice but to use un-escaped and untrusted HTML
+        // but the html used in a sandboxed iframe for protection.
+        final String iFrameHtmlContent = new StringBuilder()
+                .append("<?xml version=\"1.0\" encoding=\"UTF-8\"?>")
+                .append("<!DOCTYPE html>")
+                .append("<html class=\"")
+                .append(SafeHtmlUtil.from(currentPreferenceClasses).asString())
+                .append("\">")
+                .append("<head>")
+                .append("<meta http-equiv=\"content-type\" content=\"text/html; charset=UTF-8\">")
+                .append("<link rel=\"stylesheet\" href=\"css/app.css\" type=\"text/css\" />")
+                .append("</head>")
+                .append("<body>")
+                .append("<div class=\"max info-page markdown-container markdown ")
+                .append(SafeHtmlUtil.from(prismThemeClassName).asString())
+                .append("\">")
+                .append(markdownAsHtml)
+                .append("</div>")
+                .append("</body>")
+                .append("</html>")
+                .toString();
+
+        iFramePresenter.setSrcDoc(iFrameHtmlContent);
+    }
+
+    private String getPrismThemeClassName() {
+        final ThemeType themeType = userPreferencesManager.geCurrentThemeType();
+        switch (themeType) {
+            case DARK:
+                return "";
+            case LIGHT:
+                return PRISM_LIGHT_THEME_CLASS_NAME;
+            default:
+                throw new RuntimeException("Unexpected theme type " + themeType);
+        }
     }
 
     public void setReadOnly(final boolean readOnly) {
         this.readOnly = readOnly;
     }
 
-    public native String getHtml(final String markdown) /*-{
-        var converter = new $wnd.showdown.Converter();
-        return converter.makeHtml(markdown);
+    /**
+     * Use ShowdownJS to convert markdown text into HTML. Enable support for
+     * markdown tables.
+     */
+    public native String convertMarkdownToHtml(final String rawMarkdown) /*-{
+        // See https://showdownjs.com/docs/available-options/
+        var converter = new $wnd.showdown.Converter({
+            openLinksInNewWindow: true,
+            strikethrough: true,
+            tables: true,
+            tasklists: true,
+        });
+        var markdownAsHtml = converter.makeHtml(rawMarkdown);
+
+        // Only need to involve prism if showdown has set a language-XXX class
+        if (/language-/.test(markdownAsHtml)) {
+            // Transient div just so we can tokenise the existing html
+            var markdownDiv = $doc.createElement("div");
+            markdownDiv.innerHTML = markdownAsHtml;
+
+            // This will run the prismjs code to tokenise the content of fenced blocks which
+            // are within pre/code. It will surround the different bits of syntax with appropriately classed
+            // span tags. ShowdownJS will have set the appropriate language class on the
+            // code tag so then the prism css can style all the created spans.
+            $wnd.Prism.highlightAllUnder(markdownDiv);
+            markdownAsHtml = markdownDiv.innerHTML;
+        }
+        return markdownAsHtml;
     }-*/;
+
+
+    // --------------------------------------------------------------------------------
+
 
     public interface MarkdownEditView extends View {