+++ stroom-search/stroom-search-impl/src/main/java/stroom/search/impl/LuceneSearchProvider.java
@@ -31,8 +31,8 @@ import stroom.query.api.v2.ExpressionUtil;
 import stroom.query.api.v2.Query;
 import stroom.query.api.v2.SearchRequest;
 import stroom.query.common.v2.CoprocessorSettings;
-import stroom.query.common.v2.Coprocessors;
 import stroom.query.common.v2.CoprocessorsFactory;
+import stroom.query.common.v2.CoprocessorsImpl;
 import stroom.query.common.v2.DataStoreSettings;
 import stroom.query.common.v2.ResultStore;
 import stroom.query.common.v2.ResultStoreFactory;
@@ -66,7 +66,8 @@ public class LuceneSearchProvider implements SearchProvider {
     private final ResultStoreFactory resultStoreFactory;
     private final TaskContextFactory taskContextFactory;
     private final ExecutorProvider executorProvider;
-    private final LuceneSearchExecutor luceneSearchExecutor;
+    private final FederatedSearchExecutor federatedSearchExecutor;
+    private final LuceneNodeSearchTaskCreator luceneNodeSearchTaskCreator;
 
     @Inject
     public LuceneSearchProvider(final IndexStore indexStore,
@@ -77,7 +78,8 @@ public class LuceneSearchProvider implements SearchProvider {
                                 final ResultStoreFactory resultStoreFactory,
                                 final TaskContextFactory taskContextFactory,
                                 final ExecutorProvider executorProvider,
-                                final LuceneSearchExecutor luceneSearchExecutor) {
+                                final FederatedSearchExecutor federatedSearchExecutor,
+                                final LuceneNodeSearchTaskCreator luceneNodeSearchTaskCreator) {
         this.indexStore = indexStore;
         this.wordListProvider = wordListProvider;
         this.maxBooleanClauseCount = searchConfig.getMaxBooleanClauseCount();
@@ -86,7 +88,8 @@ public class LuceneSearchProvider implements SearchProvider {
         this.resultStoreFactory = resultStoreFactory;
         this.taskContextFactory = taskContextFactory;
         this.executorProvider = executorProvider;
-        this.luceneSearchExecutor = luceneSearchExecutor;
+        this.federatedSearchExecutor = federatedSearchExecutor;
+        this.luceneNodeSearchTaskCreator = luceneNodeSearchTaskCreator;
     }
 
     @Override
@@ -153,7 +156,7 @@ public class LuceneSearchProvider implements SearchProvider {
         // Create a handler for search results.
         final DataStoreSettings dataStoreSettings = DataStoreSettings
                 .createBasicSearchResultStoreSettings();
-        final Coprocessors coprocessors = coprocessorsFactory.create(
+        final CoprocessorsImpl coprocessors = coprocessorsFactory.create(
                 modifiedSearchRequest.getSearchRequestSource(),
                 modifiedSearchRequest.getKey(),
                 coprocessorSettingsList,
@@ -162,7 +165,7 @@ public class LuceneSearchProvider implements SearchProvider {
 
         // Create an asynchronous search task.
         final String searchName = "Search '" + modifiedSearchRequest.getKey().toString() + "'";
-        final AsyncSearchTask asyncSearchTask = new AsyncSearchTask(
+        final FederatedSearchTask federatedSearchTask = new FederatedSearchTask(
                 modifiedSearchRequest.getSearchRequestSource(),
                 modifiedSearchRequest.getKey(),
                 searchName,
@@ -177,7 +180,7 @@ public class LuceneSearchProvider implements SearchProvider {
                 coprocessors);
         resultStore.addHighlights(highlights);
 
-        luceneSearchExecutor.start(asyncSearchTask, resultStore);
+        federatedSearchExecutor.start(federatedSearchTask, resultStore, luceneNodeSearchTaskCreator);
 
         return resultStore;
     }