+++ stroom-search/stroom-search-impl/src/test/java/stroom/search/impl/TestSearchResultCreation.java
@@ -27,6 +27,7 @@ import stroom.query.api.v2.TableSettings;
 import stroom.query.common.v2.CoprocessorSettings;
 import stroom.query.common.v2.Coprocessors;
 import stroom.query.common.v2.CoprocessorsFactory;
+import stroom.query.common.v2.CoprocessorsImpl;
 import stroom.query.common.v2.DataStore;
 import stroom.query.common.v2.DataStoreFactory;
 import stroom.query.common.v2.DataStoreSettings;
@@ -121,7 +122,7 @@ class TestSearchResultCreation {
                 sizesProvider,
                 dataStoreFactory);
         final List<CoprocessorSettings> coprocessorSettings = coprocessorsFactory.createSettings(searchRequest);
-        final Coprocessors coprocessors = coprocessorsFactory.create(
+        final CoprocessorsImpl coprocessors = coprocessorsFactory.create(
                 SearchRequestSource.createBasic(),
                 queryKey,
                 coprocessorSettings,
@@ -229,7 +230,7 @@ class TestSearchResultCreation {
                 sizesProvider,
                 dataStoreFactory);
         final List<CoprocessorSettings> coprocessorSettings = coprocessorsFactory.createSettings(searchRequest);
-        final Coprocessors coprocessors = coprocessorsFactory.create(
+        final CoprocessorsImpl coprocessors = coprocessorsFactory.create(
                 SearchRequestSource.createBasic(),
                 queryKey,
                 coprocessorSettings,
@@ -242,7 +243,7 @@ class TestSearchResultCreation {
         final int[] mappings = createMappings(coprocessors.getFieldIndex());
 
         final QueryKey queryKey2 = new QueryKey(UUID.randomUUID().toString());
-        final Coprocessors coprocessors2 = coprocessorsFactory.create(
+        final CoprocessorsImpl coprocessors2 = coprocessorsFactory.create(
                 SearchRequestSource.createBasic(),
                 queryKey2,
                 coprocessorSettings,
@@ -304,7 +305,7 @@ class TestSearchResultCreation {
                 sizesProvider,
                 dataStoreFactory);
         final List<CoprocessorSettings> coprocessorSettings = coprocessorsFactory.createSettings(searchRequest);
-        final Coprocessors coprocessors = coprocessorsFactory.create(
+        final CoprocessorsImpl coprocessors = coprocessorsFactory.create(
                 SearchRequestSource.createBasic(),
                 queryKey,
                 coprocessorSettings,
@@ -317,7 +318,7 @@ class TestSearchResultCreation {
         final int[] mappings = createMappings(coprocessors.getFieldIndex());
 
         final QueryKey queryKey2 = new QueryKey(UUID.randomUUID().toString());
-        final Coprocessors coprocessors2 = coprocessorsFactory.create(
+        final CoprocessorsImpl coprocessors2 = coprocessorsFactory.create(
                 SearchRequestSource.createBasic(),
                 queryKey2,
                 coprocessorSettings,
@@ -365,7 +366,7 @@ class TestSearchResultCreation {
         validateSearchResponse(searchResponse);
     }
 
-    private void complete(final Coprocessors coprocessors) throws InterruptedException {
+    private void complete(final CoprocessorsImpl coprocessors) throws InterruptedException {
         coprocessors.getCompletionState().signalComplete();
         // Wait for the coprocessors to finish processing data.
         coprocessors.getCompletionState().awaitCompletion();
@@ -393,7 +394,7 @@ class TestSearchResultCreation {
         final QueryKey queryKey = new QueryKey(UUID.randomUUID().toString());
         final CoprocessorsFactory coprocessorsFactory = new CoprocessorsFactory(sizesProvider, dataStoreFactory);
         final List<CoprocessorSettings> coprocessorSettings = coprocessorsFactory.createSettings(searchRequest);
-        final Coprocessors coprocessors = coprocessorsFactory.create(
+        final CoprocessorsImpl coprocessors = coprocessorsFactory.create(
                 SearchRequestSource.createBasic(),
                 queryKey,
                 coprocessorSettings,
@@ -406,7 +407,7 @@ class TestSearchResultCreation {
         final int[] mappings = createMappings(coprocessors.getFieldIndex());
 
         final QueryKey queryKey2 = new QueryKey(UUID.randomUUID().toString());
-        final Coprocessors coprocessors2 = coprocessorsFactory.create(
+        final CoprocessorsImpl coprocessors2 = coprocessorsFactory.create(
                 SearchRequestSource.createBasic(),
                 queryKey2,
                 coprocessorSettings,